+++ core/src/com/gsma/rcs/platform/network/AndroidSocketConnection.java
@@ -2,6 +2,7 @@
  * Software Name : RCS IMS Stack
  *
  * Copyright (C) 2010 France Telecom S.A.
+ * Copyright (C) 2015 Sony Mobile Communications Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +15,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ * NOTE: This file has been modified by Sony Mobile Communications Inc.
+ * Modifications are licensed under the License.
  ******************************************************************************/
 
 package com.gsma.rcs.platform.network;
@@ -196,9 +200,10 @@ public class AndroidSocketConnection implements SocketConnection {
      * @param timeout Timeout in milliseconds
      * @throws IOException
      */
-    public void setSoTimeout(int timeout) throws IOException {
+    public void setSoTimeout(long timeout) throws IOException {
         if (socket != null) {
-            socket.setSoTimeout(timeout);
+            /* NOTE: External API limiting timeout that should be in type 'long' to 'int'. */
+            socket.setSoTimeout((int) timeout);
         } else {
             throw new IOException("Connection not opened");
         }
